feat: unify even without hooks
