+++ packages/libs/error-stack/src/fmt/hook.rs
@@ -440,6 +440,7 @@ mod default {
     use alloc::{format, vec, vec::Vec};
     use core::{
         any::TypeId,
+        fmt::Write,
         panic::Location,
         sync::atomic::{AtomicBool, Ordering},
     };
@@ -455,6 +456,8 @@ mod default {
     #[cfg(feature = "spantrace")]
     use tracing_error::SpanTrace;
 
+    #[cfg(feature = "pretty-print")]
+    use crate::fmt::location::LocationDisplay;
     use crate::{
         fmt::hook::{into_boxed_hook, BoxedHook, HookContext},
         Frame, Report,
@@ -494,40 +497,9 @@ mod default {
         });
     }
 
-    #[cfg(feature = "pretty-print")]
-    struct LocationDisplay<'a> {
-        location: &'a Location<'static>,
-        supports_color: bool,
-    }
-
-    #[cfg(feature = "pretty-print")]
-    impl<'a> core::fmt::Display for LocationDisplay<'a> {
-        fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-            let location = self.location;
-
-            if self.supports_color {
-                core::fmt::Display::fmt(&OwoColorize::bright_black(location), f)
-            } else {
-                f.write_fmt(format_args!("at {location}"))
-            }
-        }
-    }
-
     fn location(location: &Location<'static>, context: &mut HookContext<Location<'static>>) {
         #[cfg(feature = "pretty-print")]
-        {
-            let display = LocationDisplay {
-                location,
-                supports_color: false,
-            };
-
-            let body = display.if_supports_color(Stream::Stdout, |value| LocationDisplay {
-                location: value.location,
-                supports_color: true,
-            });
-
-            context.push_body(body.to_string());
-        }
+        context.push_body(LocationDisplay::new(location).render());
 
         #[cfg(not(feature = "pretty-print"))]
         context.push_body(format!("at {location}"));
