+++ packages/libs/error-stack/src/fmt/mod.rs
@@ -144,6 +144,8 @@
 
 #[cfg(any(feature = "std", feature = "hooks"))]
 mod hook;
+#[cfg(feature = "pretty-print")]
+mod location;
 
 use alloc::{
     borrow::ToOwned,
@@ -163,6 +165,8 @@ use core::{
 pub use hook::HookContext;
 #[cfg(any(feature = "std", feature = "hooks"))]
 pub(crate) use hook::{install_builtin_hooks, Format, Hooks};
+#[cfg(all(not(any(feature = "std", feature = "hooks")), feature = "pretty-print"))]
+use location::LocationDisplay;
 #[cfg(feature = "pretty-print")]
 use owo_colors::{OwoColorize, Stream, Style as OwOStyle};
 
@@ -703,13 +707,7 @@ fn debug_attachments_invoke<'a>(
             #[cfg(all(not(any(feature = "std", feature = "hooks")), feature = "pretty-print"))]
             FrameKind::Attachment(AttachmentKind::Opaque(_)) => frame
                 .downcast_ref::<core::panic::Location<'static>>()
-                .map(|location| {
-                    vec![
-                        location
-                            .if_supports_color(Stream::Stdout, OwoColorize::bright_black)
-                            .to_string(),
-                    ]
-                }),
+                .map(|location| vec![LocationDisplay::new(location).render()]),
             #[cfg(all(
                 not(any(feature = "std", feature = "hooks")),
                 not(feature = "pretty-print")
