+++ packages/libs/error-stack/src/lib.rs
@@ -464,6 +464,7 @@
 )]
 
 extern crate alloc;
+extern crate core;
 
 pub mod future;
 pub mod iter;
